Refactor: move code around so its more organized
